Moves Framework-related functions from `internal/flex` to `internal/framework/flex`
